+++ src/main/java/shop/itbook/itbookshop/category/exception/AlreadyAddedCategoryNameException.java
@@ -1,4 +1,4 @@
-package shop.itbook.itbookshop.category.service.impl;
+package shop.itbook.itbookshop.category.exception;
 
 /**
  * @author 최겸준
@@ -8,7 +8,7 @@ public class AlreadyAddedCategoryNameException extends RuntimeException {
 
     public static final String MESSAGE = "해당 메인카테고리의 자식카테고리에 이미 같은 이름의 카테고리가 존재합니다.";
 
-    AlreadyAddedCategoryNameException() {
+    public AlreadyAddedCategoryNameException() {
         super(MESSAGE);
     }
 }
